// frontend/src/app/model/family-member.ts
export class FamilyMember {
  [k: string]: any;
  _id: number | string = '';
  first_name: string = '';
  last_name: string = '';
  email: string = '';
  role: number = 1 | 2 | 3;
  password: string = '';
  nickname: string = '';
  kinshipDegree: string = '';
  favouriteGenre: string = '';
  favouriteMovie: string = '';
}